requirements.txt
Django==2.2.9
django-discover-runner==1.0
django-role-permissions==3.1.1
Pillow==7.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
